requirements.txt
google-generativeai
pandas
folium
tqdm
markdown
pygbif